+++ include/ecs2/components/transform_component.h
@@ -26,31 +26,37 @@ class TransformComponent : public IComponent
 
         float angle() const { return _angle; }
 
-        Vector2D<int> direction() const { return _direction; }
+        Vector2D<float> direction() const { return _direction; }
 
-        Vector2D<int> position() const { return _position; }
+        Vector2D<float> position() const { return _position; }
 
-        Vector2D<int> velocity() const { return _velocity; }
+        Vector2D<float> velocity() const { return _velocity; }
 
         void set_angle(float angle) { _angle = angle; }
 
-        void set_direction(Vector2D<int> direction)
+        void set_direction(Vector2D<float> direction)
         {
             _direction = direction;
         }
 
-        void set_position(int x, int y) { _position = {x, y}; }
+        void set_position(int x, int y)
+        {
+            _position = {static_cast<float>(x), static_cast<float>(y)};
+        }
 
-        void set_velocity(const Vector2D<int>& vel) { _velocity = vel; }
+        void set_velocity(const Vector2D<float>& vel)
+        {
+            _velocity = vel;
+        }
 
         void set_x_velocity(int x_vel)
         {
-            _velocity = {x_vel, _velocity.y()};
+            _velocity = {static_cast<float>(x_vel), _velocity.y()};
         }
 
         void set_y_velocity(int y_vel)
         {
-            _velocity = {_velocity.x(), y_vel};
+            _velocity = {_velocity.x(), static_cast<float>(y_vel)};
         }
 
         void update(const SDL_Event& /*e*/) override
@@ -66,9 +72,9 @@ class TransformComponent : public IComponent
         int _speed{3};
         float _angle{.0};
 
-        Vector2D<int> _position;
-        Vector2D<int> _velocity;
-        Vector2D<int> _direction;
+        Vector2D<float> _position;
+        Vector2D<float> _velocity;
+        Vector2D<float> _direction;
 };
 
 using TransformComponentSPtr = std::shared_ptr<TransformComponent>;
